agent/requirements.txt
strands-agents
bedrock-agentcore
httpx